asn1 tests: Let ts:run() build the tests
Only release the source code and an Emakefile, so that ts:run()
can build the tests just before running them.

The main advantage of building the tests suite like that is
that the turn-around time for developing new test cases are
much faster: just edit the appropriate file and do ts:run().