Implement encoding of fragmented OCTET STRINGs
Decoding of a fragmented OCTET STRING with the size constrained to a
single value is implemented, but encoding is not implemented, and
decoding is not tested in any test case.

There is no urgent need to support encoding of fragmented OCTET STRINGs,
but not being able to encode such values complicates testing of decoding.
For example, eif we add the following type to PrimStrings.asn1 in the test
suite:

  FixedOs65536 ::= OCTET STRING (SIZE (65536))

then asn1ct:test('PrimStrings') will fail because this value cannot
be encoded.