BER run-time: Refactor decoding of string data types
In the BER run-time code (asn1rtt_ber), there is a general function
for decoding a string:

  decode_restricted_string(Tlv, Range, StringType, TagsIn)

But the StringType argument is not important. It is only used
internally in asn1rtt_ber to distinguish universal strings and BMP
strings from other strings. By slightly refactoring the string
decoding code, we can eliminate the StringType argument and a
subsequent run-time test on the string type. That will slightly reduce
code size, slightly increase speed, and eliminate Dialyzer warnings.