Eliminate the {notype,_} return value from asn1ct_gen:type/1
The last clause in asn1ct_gen:type/1 does a catched call to type2/1.
If the type2/1 fails {notype,X} is returned.

Since the body of type2/1 essentially is:

  case lists:member(X, [...]) of
    true ->
      {primitive,bif};
    false ->
      case lists:member(X, [...]) of
         true ->
           {constructed,bif};
         false ->
           {undefined,user}
      end
   end

there is no way that type2/1 can fail. Therefore, we can eliminate
the catch and put the body of type2/1 into the last clause of
type/1. We can also eliminate the code in the callers of type/1
that match {notype,X}.