Clean up checking of values for ENUMERATEDs
Unify the code for checking an enumeration value named in a
DEFAULT and in an ENUMERATED value. There is no need to handle
those cases differently. That also will also make sure that
the following works:

  E ::= ENUMERATED { x, ..., y }
  e E ::= x

(Extensible ENUMERATEDs were not handled when defining values.)

Always generate an error when an unknown enumeration value is
given (used in a DEFAULT, a message would be printed, but the
compilation would succeed). Also make sure that we always include
the line number for the incorrect enumeration.

Write a new test case and remove the extremely rudimentary
value_bad_enum_test/1 test case.